Implement cosim purely in a HarnessBindeR
