+++ generators/chipyard/src/main/scala/DigitalTop.scala
@@ -34,7 +34,6 @@ class DigitalTop(implicit p: Parameters) extends ChipyardSystem
   with chipyard.clocking.HasChipyardPRCI // Use Chipyard reset/clock distribution
   with fftgenerator.CanHavePeripheryFFT // Enables optionally having an MMIO-based FFT block
   with constellation.soc.CanHaveGlobalNoC // Support instantiating a global NoC interconnect
-  with chipyard.CanHaveSpikeCosim // Support instantiating spike-based co-simulation
 {
   override lazy val module = new DigitalTopModule(this)
 }
