+++ generators/chipyard/src/main/scala/config/fragments/TileFragments.scala
@@ -9,7 +9,6 @@ import freechips.rocketchip.rocket.{RocketCoreParams, MulDivParams, DCacheParams
 
 import boom.common.{BoomTileAttachParams}
 import cva6.{CVA6TileAttachParams}
-import chipyard.{SpikeCosimKey}
 import testchipip._
 
 class WithL2TLBs(entries: Int) extends Config((site, here, up) => {
@@ -80,6 +79,3 @@ class WithRocketDCacheScratchpad extends Config((site, here, up) => {
   }
 })
 
-class EnableSpikeCosim extends Config((site, here, up) => {
-  case SpikeCosimKey => true
-})
